+++ changelog.md
@@ -1,6 +1,8 @@
-- switch from haskell-src-exts to ghc-lib-parser.  This fixes a bunch of
+- Switch from haskell-src-exts to ghc-lib-parser.  This fixes a bunch of
 parsing bugs.  Along the way I fixed a bug where locally bound names were
 misinterpreted as unqualified names.
+
+- Detect and abort on CPP in the import block, instead of silently deleting it.
 
 ### 2.3.0